bankapp/requirements.txt
Django==1.10.6
djangorestframework==3.6.2
psycopg2==2.7.1
pep8==1.7.0
django-wkhtmltopdf==3.1.0